+++ packages/core/src/components/cv-breadcrumb/cv-breadcrumb-item.vue
@@ -11,7 +11,7 @@
 </template>
 
 <script>
-import carbonPrefixMixin from '../../mixins/carbon-prefix-mixin';
+import { carbonPrefixMixin } from '../../mixins';
 export default {
   name: 'CvBreadcrumbItem',
   mixins: [carbonPrefixMixin],