+++ packages/core/src/components/cv-inline-notification/cv-inline-notification.vue
@@ -46,13 +46,12 @@
 </template>
 
 <script>
-import notificationMixin from '../../mixins/notification-mixin';
+import { notificationMixin, carbonPrefixMixin } from '../../mixins';
 import ErrorFilled20 from '@carbon/icons-vue/es/error--filled/20';
 import CheckmarkFilled20 from '@carbon/icons-vue/es/checkmark--filled/20';
 import WarningFilled20 from '@carbon/icons-vue/es/warning--filled/20';
 import Close20 from '@carbon/icons-vue/es/close/20';
 import InformationFilled20 from '@carbon/icons-vue/es/information--filled/20';
-import carbonPrefixMixin from '../../mixins/carbon-prefix-mixin';
 
 export default {
   name: 'CvInlineNotification',