+++ packages/core/src/components/cv-number-input/cv-number-input-skeleton.vue
@@ -6,7 +6,7 @@
 </template>
 
 <script>
-import carbonPrefixMixin from '../../mixins/carbon-prefix-mixin';
+import { carbonPrefixMixin } from '../../mixins';
 export default {
   name: 'CvNumberInputSkeleton',
   mixins: [carbonPrefixMixin],