+++ packages/core/src/components/cv-skeleton-text/cv-skeleton-text.vue
@@ -10,7 +10,7 @@
 </template>
 
 <script>
-import carbonPrefixMixin from '../../mixins/carbon-prefix-mixin';
+import { carbonPrefixMixin } from '../../mixins';
 export default {
   name: 'CvSkeletonText',
   mixins: [carbonPrefixMixin],