+++ packages/core/src/components/cv-tag/cv-tag.vue
@@ -8,7 +8,6 @@
         [`${carbonPrefix}--tag--disabled`]: disabled,
       },
     ]"
-    :tabindex="!disabled ? 0 : undefined"
     role="listitem"
     :title="title"
     @keydown.enter.stop.prevent="$emit('remove')"
@@ -30,7 +29,7 @@
 
 <script>
 import Close16 from '@carbon/icons-vue/es/close/16';
-import carbonPrefixMixin from '../../mixins/carbon-prefix-mixin';
+import { carbonPrefixMixin } from '../../mixins';
 
 const tagKinds = [
   'red',