+++ packages/core/src/components/cv-tile/_cv-tile-clickable.vue
@@ -13,12 +13,10 @@
 </template>
 
 <script>
-import LinkMixin from '../../mixins/link-mixin';
-import carbonPrefixMixin from '../../mixins/carbon-prefix-mixin';
-import methodsMixin from '../../mixins/methods-mixin';
+import { linkMixin, carbonPrefixMixin, methodsMixin } from '../../mixins';
 
 export default {
   name: 'CvTileClickable',
-  mixins: [LinkMixin, carbonPrefixMixin, methodsMixin({ target: ['blur', 'focus'] })],
+  mixins: [linkMixin, carbonPrefixMixin, methodsMixin({ target: ['blur', 'focus'] })],
 };
 </script>